pyproject.toml
[project]
name = "video-extractor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "Rong Zhiwei", email = "18845728185@163.com" }]
requires-python = ">=3.12"
dependencies = ["openai-whisper>=20240930", "torch>=2.6.0", "yt-dlp>=2025.3.31"]

[project.scripts]
video-extract = "video_extractor:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources]
torch = [{ index = "pytorch-cu126", marker = "platform_system == 'Windows'" }]

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true

[dependency-groups]
dev = ["ipdb>=0.13.13"]
